+++ TP_Pokemon_Console/plantpokemon.h
@@ -6,7 +6,7 @@
 class PlantPokemon : public Pokemon
 {
 public:
-    PlantPokemon(const std::string &itsName, int itsSize, int itsWeight, int itsHealthPoint, int itsStrengthPower);
+    PlantPokemon(const std::string &itsName, float itsSize, float itsWeight, int itsHealthPoint, int itsStrengthPower);
     // Pokemon interface
     void calculateSpeed();
     std::string getDescription();
